fix include comments
